+++ js/constants.js
@@ -1,0 +1,15 @@
+const DEFAULT_MIN_STEP_VALUE = 1;
+const DEFAULT_MAX_STEP_VALUE = 99;
+const DEFAULT_NUMBER_FLOAT_VALUE = 0;
+const X_MIN_COORDINATE = 35.65000;
+const X_MAX_COORDINATE = 35.70000;
+const Y_MIN_COORDINATE = 139.70000;
+const Y_MAX_COORDINATE = 139.80000;
+const COORDINATE_NUMBER_FLOAT_VALUE = 5;
+const MIN_PRICE = 500;
+const MAX_PRICE = 10000;
+const MAX_AMOUNT_USERS = 8;
+const MIN_AMOUNT_USERS = 1;
+const ARRAYDATAAMOUNT = 10;
+
+export {DEFAULT_MIN_STEP_VALUE,DEFAULT_MAX_STEP_VALUE,DEFAULT_NUMBER_FLOAT_VALUE,X_MIN_COORDINATE,X_MAX_COORDINATE,Y_MIN_COORDINATE,Y_MAX_COORDINATE,COORDINATE_NUMBER_FLOAT_VALUE,MIN_PRICE,MAX_PRICE,MAX_AMOUNT_USERS,MIN_AMOUNT_USERS,ARRAYDATAAMOUNT};
