+++ js/server.js
@@ -1,0 +1,56 @@
+const TITLES = [
+  'Дворец из коробки',
+  'Дом с бассейном',
+  'Просторное Бунгало',
+  'Отель Marriot Tokyo',
+  'Квартира с видом на площадь',
+];
+const DESCRIPTIONS = [
+  'Замечательный дворец в старинном центре города. Только для тех кто может себе позволить дворец.',
+  'Прекрасный дом с большой территорией, красивой большой терассой а так же большим бассейном. РАсположеный в  10 минутах от центра Токио.',
+  'Очень уютной бунгало на краю города, в дали от суеты мегаполиса',
+  'Хорошо всем известный мировой бренд отель Marriot.',
+  'Уютная квартира в самом центре Токио с красивым видом из окна!',
+];
+const HOUSES_TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+const TIMES = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+const ROOMS_AMOUNT = [
+  1,
+  2,
+  3,
+  50,
+  100,
+];
+
+const GUEST_AMOUNT = [
+  3,
+  2,
+  1,
+  0,
+];
+
+const PHOTOS_URLS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+export {TITLES,DESCRIPTIONS,HOUSES_TYPES,FEATURES,TIMES,ROOMS_AMOUNT,GUEST_AMOUNT,PHOTOS_URLS};
